Extract Korean address logic into its own builder
